http: another cut at fixing HTTP framing
For keep-alive connections it's neither sufficient to cut at the end of
header (since the rest of the response would leak into next request), nor
just fan out the current SSL frame (since the SSL framing is effectively
undefined and we don't know if we're done yet and with non-blocking reads
we're likely to just terminate too early).

So, let's just obey the Content-Length header. If it's missing (chunked
encoding or standards violation) then read until the EOF, but still
succeed if the header was complete. This way we're hopefully on the safe
side.