Add support for realm authentication
The official fortinet client accepts a connection string of the form:
https://<vpnhost>:<port>/<realm>
Most VPN setups do not use the realm, so it's empty. It's only used
to provide multiple VPNs with different configuration on the same gateway.

This commit adds a --realm option to specify an authentication realm
to authenticate to. It defaults to empty.