Perform 2FA auth with zero-length tokeninfo
Previously, we assumed that the tokeninfo parameter must be present and
contain data if we are to perform two-factor authentication. Based on
GitHub issue #41, in some cases, the tokeninfo parameter will be present
but contain an empty string.

This commit removes the check for string length, as it was not needed.