io.c: fix core cause of openfortivpn is locked when spawning pppd has failed.
This commit reverts the temporary fix from #132 (commit 158f489)
and instead replaces `GCD semaphores` with `Mach Semaphores`,
following the lines of http://hg.openjdk.java.net/jdk8u/jdk8u/hotspot/file/8b16790cd73a/src/os/bsd/vm/os_bsd.cpp.

By doing so, this commit is also a solution for issue #105(GCD semaphores might be unsafe when used in signal handlers):
On macOS `named POSIX semaphores`, which are known to be async-signal-safe, are implemented upon `Mach Semaphores`.

An analysis of the use of semaphores in openfortivpn has lead to the conclusion
that `Mach Semaphores` are just as good as `unnamed POSIX semaphores`
which unfortunately are not implemented in macOS.

※ About `Mach Semaphores`
- https://books.google.de/books?id=K8vUkpOXhN4C&pg=PA1219&lpg=PA1210&focus=viewport&dq=semaphore_create
- https://developer.apple.com/library/content/documentation/Darwin/Conceptual/KernelProgramming/synchronization/synchronization.html