build: avoid evaluating $sysconfidir on configure time
${prefix} doesn't make sense there. At least with current autotools
version it is set to "NONE" which is really not what we want:

  [lkundrak@belphegor openfortivpn]$ ./openfortivpn localhost:10443
  WARN:   Could not load config file "NONE/etc/openfortivpn/config" (No such file or directory).

Also, the expansion of SYSCONFDIR is done in the Makefile and then
manual is templated using sed, which is what seems to be the correct way
to do that with autotools -- as opposed to letting config.status
templating it. Othewise expansion of ${sysconfig} gets peculiar.