clean up config initialization and error messages during parsing options (#167)
initialization took place in a macro, except for a few variables of the config. This patch cleans that up and moves all initialization into the macro. There were also a few grammar mistakes in the error messages that may printed out when wrong values were supplied for the command line options