Fix Coverity defect
CID 182341: String not null terminated (STRING_NULL)
	string_null_argument: Function fread does not terminate
	string *buffer.

This is a problem in the edge case where /etc/resolv.conf is empty
or does not contain a new line '\n' character. Then strtok() isn't
stopped by either '\n' or '\0' and fells through.

Additionally we print a specific error message and bail out if
/etc/resolv.conf is empty.