Does /usr/sbin/pppd exist?
If /usr/sbin/pppd does not exist, print a specific message. When users
build from source, it happens often enough that PPP is not installed.

The existing frprintf(stderr, ...) call just after execvp() currently
doesn't work, probably because of the prior forkpty() call. If we get
it to work someday, we can move this after the failed call to execvp().
On the other hand a very specific message in case /usr/sbin/pppd is
missing makes sense, as this occurs quite often.

Because of limitations of C const-correctness, the execpv signature is:
	char *const argv[]
but execvp() really doesn't change its arguments. Therefore it is valid
to pass an array of:
	const char *const
then casti the array back to:
	char *const *
just before calling execvp().