call aclocal from autogen.sh only if it exists
the automake manual discourages the use of aclocal because on the long
term it is supposed to go away. However, with automake 1.15 it still
needs to be called (otherwise even built-in macros are not working).
With the current commit a smooth transition towards future autotools
that don't provide aclocal anymore should be possible