If the OTP is specified in the configuration, use it for 2FA (#255)
* If the OTP is specified in the configuration, use it for 2FA

Assume the second factor is the one-time password. The distinction
between a mechanism OTP is requested from how 2FA token is requested
doesn't provide much clue into what's the distinction. Assume they're
the same thing, not used together.

This allows the NetworkManager (that sets the otp key in config) to work
with 2FA setups.

* apply url_encode on otp when used as 2fa input