Rework OpenSSL library detection
Currently, X509_check_host function is always marked as missing, since
AC_CHECK_FUNCS is missing linker flags for libssl and libcrypto.

Culprit for this is the PKG_CHECK_MODULES call that does not add
libraries to the link arguments that configure uses to create function
tests, which in turn makes function check fail.

To remedy the situation, this commit replaces PKG_CHECK_MODULES with
two AC_CHECK_LIB calls (one for libcrypto and one for libssl) that
properly populate configure linker flags.