Rationalize DNS options
* The --set-dns/set-dns option instructs either of ppp/openfortivpn
to get DNS settings from peer ppp/FortiGate server and rewrite
/etc/resolv.conf.
* The --pppd-no-peerdns/pppd-use-peerdns switches the above
functionality from ppp to openfortivpn. This option is marked
as deprecated.

The difference with the previous behaviour is that:
* --set-dns=0/set-dns=0 disables DNS handling in both ppp and
openfortivpn instead of openfortivpn only. A single option is now
sufficient to avoid messing with /etc/resolv.conf.
* --pppd-no-peerdns/pppd-use-peerdns=0 switches DNS handling from
pppd to openfortivpn. It didn't make sense to have both programs
compete to modify /etc/resolv.conf, did it? Feels much better this
way.