read_password: do not unset ICANON bytes on terminal
we do not actually need to clear ICANON, the terminal will handle these
control sequences for us and only feed us the whole line once it is
ready.

This avoids having to process the letters one at a time, for example
incorrectly including non-printable characters in the password string
when the user presses backspace instead of deleting the last character:
with this change, backspace, ^H and ^U all work as usual.