fix overrides in configure.ac
command line arguments (--with-... and --enable-...) are processed first,
so it is clearer to move their definition more to the top of configure.ac.

existence of files is processed later, so we have to put additional if-
statements around so that the already processed overrides are not touched.

we may not assign an empty value to the with_... and enable_... shell variables
in configure.ac, because that would also override the values from command line