cipher-list option was always ignored
the cipher list option has always been considered only inside a branch
which handles the case when it is not set. So, the default setting
did always win. When the option was set in the config or on the command line
there was no code to handle it, so the default from openssl has been
taken in that case