update documentation (#419)
update help message and man page:
- indicate the fact that the tcp port may be omitted now (defaults to https)
- better explanation of --set-dns and --pppd-no-peerdns options
- mention --ppp-system in man page
- mention in man page that pppd/ppp options are available only if compiled in
- update example config file in the man page
- minor fixes of typos etc.
- update date of man page
Additionally, the length of the help string makes it necessary to move some of the options into a separate string. Since we have the ppp/pppd options already defined separately, it is the easiest to pass them as an extra string to printf between the other options and the config file example.