tunnel: pass our stderr to the pppd slave
While forkpty() does a nice job of convenently setting up a pty for the
pppd, it goes a bit too far for our purposes by also sending stderr to
the pseudoterminal.

This causes confusion in the event pppd (or its plugin, or us, for that
matter) try to log anything to stderr. Let's restore the parent's stderr
after he fork.