userinput: add capability of spawning a helper to get a secret
This adds the --pinentry option that allows passing a name of a program
that can supply a password. This is going to be useful in order to relay
the password requestes to the NetworkManager plugin.

The protocol used is compatible with what GnuPG's pinentry program uses
(documented in pinentry GNU Info manual), so that this could be tested
or used without the NetworkManager plugin.