correctly shift masks for cidr notation on MAC (#471)
when parsing routing entries with cidr notation, the calculated
mask did not respect the reverse byte order. This has lead to the
result that ipv4_get_route has returned the wrong route in some
cases. We have to shuffle the bytes of the netmask in order to
achieve the correct  byte order