Always define RESOLVCONF_USAGE / RESOLVCONF_HELP
While either HAVE_USR_SBIN_PPPD or HAVE_USR_SBIN_PPP *must* be defined,
HAVE_RESOLVCONF might not be defined - typically when resolvconf is not
available at build-time. RESOLVCONF_USAGE and RESOLVCONF_HELP should always
be defined, if needed as empty strings, so as to not break subsequent
printf() calls.