Attempt to warn of outdated FortiGate appliances
FortiGate appliances that are unable to provide configuration
in XML format are probably running FortiOS 4 or earlier which
has been umaintained and vulnerable for some time now.

Warn the user that the FortiGate appliance is outadated and that
it might not be compatible with recent OpenSSL 1.1.1 libraries
with weak ciphers disabled - we've had a case on Ubuntu 20.04 for
example.