Extend OTP support to also allow using FTM push
If the server claims it's supported (from the initial login response);
send the request for a push notification.

If an OTP is specified using `--otp=` or through the configuration
file, the given OTP will be used, no matter what the server initially
responded with.

Use `--no-ftm-push` to disable this altogether.