Hopefully reach higher speeds on macOS
While on Linux the speed option does not seem to be taken into account,
on macOS it appears to be throttling the VPN connection. This Baud rate
of 230400 has been reported to achieve ×10 speeds over 115200. I have no
explanation but since it works...

According to the pppd man page, Linux and other system "only support the
commonly-used baud rates".