Avoid spurious warning on macOS
On macOS, SO_SNDBUF and SO_RCVBUF are defined in <sys/socket.h> but
getsockopt() fails for these values with:
	Protocol not available

I guess this is expected on macOS, although the GETSOCKOPT(2) man page
does not document that. In any case the behaviour is common enough to
avoid the warning in the log.

Fixes #822.