Properly escape non-ASCII characters
Until now UTF-8 encoded character 'à' was escaped to:
	%FFFFFFC3%FFFFFFA9
but it should be escaped to:
	%C3%A9

The nice thing with the current uri_unescape() function is that it
works with both mappings,so we should be able to handle both mappings
transparently and avoid compatibility issues.

Note that we should make sure strings are encoded to UTF-8 before
escaping them. However, I won't enforce that here because:
1. UTF-8 should already be the default encoding on supported platforms,
2. changing the mapping might cause compatibility issues,
3. it's OK if escaping and unescaping are consistently wrong.