Attempt to release allocated resources
Properly reset variables after releasing resources.

The following resources are allocated in this order in ssl_connect():
	ssl_socket
	ssl_context
	ssl_handle
Release them in the reverse order in ssl_disconnect() and in case of
error in ssl_connect():
	ssl_handle
	ssl_context
	ssl_socket

I have not fixed cases where tunnel->ssl_socket is released, but
tunnel->ssl_context and tunnel->ssl_handle are left behind. This seems
very wrong, but fixing it breaks --persistent.