Fix CodeQL false positive
The length of char env_var[24] is enough to hold:
"VPN_ROUTE_GATEWAY_" + "65535" + '\0'

The key is that the value of tunnel->ipv4.split_routes is between 0
and MAX_SPLIT_ROUTES, which is 65535 and hence cannot be longer than
5 characters.

CodeQL doesn't see that and believes that tunnel->ipv4.split_routes
may range from -2147483646 to 2147483646, hence print as 11 characters.

That's why CodeQL suggests to increase the size of env_var from 24 to 30.
But that a false positive of course.

Attempt to work around it by adding an assert().