Improve error handling in #952
The getline() man page is misleading, instead use the POSIX documentation:
	If an error occurs, the error indicator for the stream shall be set,
	and the function shall return -1 and set errno to indicate the error.

So it appears more appropriate to use ferror() to detect errors. If you
insist on using errno to detect errors, you need to reset errno before
calling getline(), as getline() will set errno in case of error, but
will not reset errno in the absence of error.