SSO/SAML/2FA and external browser
A few notes on how to proceed in case SSO/SAML/2FA require user
interaction in web pages containing JavaScript. The best course of
action is to use external software that spawns a real web browser
to retrieve the session cookie. Of course that requires a graphic
session, unlike openfortivpn.