Don't use backticks in manpage
They are not rendered nicely, so convert most of them to italics. The
URI looking ones are converted to Courier instead. This is a workaround
for the upcoming automated conversion to the wiki page. Courier will be
converted to code spans which will not be affected by the autolink
features of the Github Wiki. Which would create broken links for them.